backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
faiss-cpu>=1.7.3
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
numpy>=1.24.0
google-cloud-aiplatform>=1.25.0
prometheus-client>=0.16.0
python-dotenv>=1.0.0
aiohttp>=3.8.4
langchain>=0.0.300
schedule>=1.2.0
python-dateutil>=2.8.2 